As of Build [v0.0.3.2] Updated DB to use UserId and not name this will allow user to change names if they want

+++ app/src/main/java/com/town/small/brewtopia/AppSettings/AppSettings.java
@@ -26,7 +26,7 @@ public class AppSettings extends ActionBarActivity {
     private ListView settingsListView;
     private List<AppSettingsSchema> settingsList;
     private DataBaseManager dbManager;
-    private int userId;
+    private long userId;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
